backend/requirements.txt
# Web framework
fastapi==0.111.1
uvicorn[standard]==0.30.1

# Database – async PostgreSQL (Neon)
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2

# Pydantic v2
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.9

# AWS S3 (file storage)
boto3==1.34.144

# HTTP client (WhatsApp API calls)
httpx==0.27.0

# Background tasks
celery[redis]==5.4.0
redis==5.0.7

# Image processing
pillow==10.4.0

# Utilities
python-dotenv==1.0.1
